removed horizontal product scrolling as per user testing
