+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/rebekahdejong/Documents/web-dev/arcane-website/src/Components/TextilesArea/index.js":"1","/Users/rebekahdejong/Documents/web-dev/arcane-website/src/Elements/TextilesAreaWrap/index.js":"2","/Users/rebekahdejong/Documents/web-dev/arcane-website/src/Components/ProductsReel/index.js":"3"},{"size":441,"mtime":1626263661855,"results":"4","hashOfConfig":"5"},{"size":2251,"mtime":1626262985579,"results":"6","hashOfConfig":"5"},{"size":2096,"mtime":1626263549299,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ep22wy",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/rebekahdejong/Documents/web-dev/arcane-website/src/Components/TextilesArea/index.js",[],"/Users/rebekahdejong/Documents/web-dev/arcane-website/src/Elements/TextilesAreaWrap/index.js",[],"/Users/rebekahdejong/Documents/web-dev/arcane-website/src/Components/ProductsReel/index.js",[]]
